fix: id should be auto-increment

+++ db/migration/20230124132848_create_articles.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "articles" (
-  "id" int PRIMARY KEY,
+  "id" bigserial PRIMARY KEY,
   "title" varchar,
   "content" text,
   "created_at" timestamptz NOT NULL DEFAULT (now()),
